backend/requirements.txt
flask-restx==1.1.0
flask_cors==4.0.0
requests==2.31.0
sseclient==0.0.27
